Feat/input verification

+++ backend/src/main/java/it/chalmers/gamma/controller/admin/AuthorityAdminController.java
@@ -17,17 +17,21 @@ import it.chalmers.gamma.response.GetAllAuthoritiesResponse;
 import it.chalmers.gamma.response.GetAllAuthorityLevelsResponse;
 import it.chalmers.gamma.response.GetAuthorityResponse;
 import it.chalmers.gamma.response.GroupDoesNotExistResponse;
-import it.chalmers.gamma.response.MissingRequiredFieldResponse;
+import it.chalmers.gamma.response.InputValidationFailedResponse;
 import it.chalmers.gamma.response.PostDoesNotExistResponse;
 import it.chalmers.gamma.service.AuthorityLevelService;
 import it.chalmers.gamma.service.AuthorityService;
 import it.chalmers.gamma.service.FKITSuperGroupService;
 import it.chalmers.gamma.service.PostService;
+import it.chalmers.gamma.util.InputValidationUtils;
 
 import java.util.List;
 import java.util.UUID;
 
+import javax.validation.Valid;
+
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -55,7 +59,10 @@ public final class AuthorityAdminController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<String> addAuthority(@RequestBody AuthorizationRequest request) {
+    public ResponseEntity<String> addAuthority(@Valid @RequestBody AuthorizationRequest request, BindingResult result) {
+        if (result.hasErrors()) {
+            throw new InputValidationFailedResponse(InputValidationUtils.getErrorMessages(result.getAllErrors()));
+        }
         Post post = this.postService.getPost(UUID.fromString(request.getPost()));
         if (post == null) {
             throw new PostDoesNotExistResponse();
@@ -77,6 +84,9 @@ public final class AuthorityAdminController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<String> removeAuthorization(@PathVariable("id") String id) {
+        if (!this.authorityService.authorityExists(UUID.fromString(id))) {
+            throw new AuthorityNotFoundResponse();
+        }
         this.authorityService.removeAuthority(UUID.fromString(id));
         return new AuthorityRemovedResponse();
     }
@@ -89,9 +99,10 @@ public final class AuthorityAdminController {
 
     // BELOW THIS SHOULD MAYBE BE MOVED TO A DIFFERENT FILE
     @RequestMapping(value = "/level", method = RequestMethod.POST)
-    public ResponseEntity<String> addAuthorityLevel(@RequestBody AuthorizationLevelRequest request) {
-        if (request.getAuthorityLevel() == null) {
-            throw new MissingRequiredFieldResponse("authorityLevel");
+    public ResponseEntity<String> addAuthorityLevel(@Valid @RequestBody AuthorizationLevelRequest request,
+                                                    BindingResult result) {
+        if (result.hasErrors()) {
+            throw new InputValidationFailedResponse(InputValidationUtils.getErrorMessages(result.getAllErrors()));
         }
         if (this.authorityLevelService.authorityLevelExists(request.getAuthorityLevel())) {
             throw new AuthorityLevelAlreadyExists();
@@ -108,6 +119,9 @@ public final class AuthorityAdminController {
 
     @RequestMapping(value = "/level/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<String> removeAuthorityLevel(@PathVariable("id") String id) {
+        if (this.authorityLevelService.authorityLevelExists(UUID.fromString(id))) {
+            throw new AuthorityNotFoundResponse();
+        }
         this.authorityLevelService.removeAuthorityLevel(UUID.fromString(id));
         return new AuthorityLevelRemovedResponse();
     }
