+++ backend/src/main/java/it/chalmers/gamma/controller/admin/WebsiteAdminController.java
@@ -5,17 +5,21 @@ import it.chalmers.gamma.requests.CreateWebsiteRequest;
 import it.chalmers.gamma.response.EditedWebsiteResponse;
 import it.chalmers.gamma.response.GetAllWebsitesResponse;
 import it.chalmers.gamma.response.GetWebsiteResponse;
-import it.chalmers.gamma.response.MissingRequiredFieldResponse;
+import it.chalmers.gamma.response.InputValidationFailedResponse;
 import it.chalmers.gamma.response.WebsiteAddedResponse;
 import it.chalmers.gamma.response.WebsiteDeletedResponse;
 import it.chalmers.gamma.response.WebsiteNotFoundResponse;
 import it.chalmers.gamma.service.GroupWebsiteService;
 import it.chalmers.gamma.service.UserWebsiteService;
 import it.chalmers.gamma.service.WebsiteService;
+import it.chalmers.gamma.util.InputValidationUtils;
 
 import java.util.List;
+import java.util.UUID;
+import javax.validation.Valid;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -42,9 +46,9 @@ public final class WebsiteAdminController {
 
 
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<String> addWebsite(@RequestBody CreateWebsiteRequest request) {
-        if (request.getName() == null) {
-            throw new MissingRequiredFieldResponse("name");
+    public ResponseEntity<String> addWebsite(@Valid @RequestBody CreateWebsiteRequest request, BindingResult result) {
+        if (result.hasErrors()) {
+            throw new InputValidationFailedResponse(InputValidationUtils.getErrorMessages(result.getAllErrors()));
         }
         this.websiteService.addPossibleWebsite(request.getName(), request.getPrettyName());
         return new WebsiteAddedResponse();
@@ -73,6 +77,9 @@ public final class WebsiteAdminController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<String> deleteWebsite(@PathVariable("id") String id) {
+        if (!this.websiteService.websiteExists(UUID.fromString(id))) {
+            throw new WebsiteNotFoundResponse();
+        }
         this.groupWebsiteService.deleteGroupWebsiteByWebsite(this.websiteService.getWebsiteById(id));
         this.userWebsiteService.deleteUserWebsiteByWebsite(this.websiteService.getWebsiteById(id));
         this.websiteService.deleteWebsite(id);
